+++ src/solver/mod.rs
@@ -1,7 +1,9 @@
 pub mod direct;
 mod direct_specialized;
 pub mod periodic_direct;
+mod direct_specialized_debug;
 
 pub use direct::*;
 pub use direct_specialized::*;
+pub use direct_specialized_debug::*;
 pub use periodic_direct::*;
